Rename team name folder
